+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2024-02-01
+### Added
+- `@rarimo/rarime`:
+  - Migration to a Compose DB model with increased Verifiable Credential size;
+  - Compose DB model preparation script;
+
 ## [2.0.1] - 2024-01-24
 ### Changed
 - `@rarimo/rarime-connector` - set strict snap version requirement
@@ -150,8 +156,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Implemented `@rarimo/rarime-connector` and `@rarimo/rarime` packages
 
-[Unreleased]: https://github.com/rarimo/rarime/compare/2.0.1...HEAD
-[2.0.0]: https://github.com/rarimo/rarime/compare/2.0.0...2.0.1
+[Unreleased]: https://github.com/rarimo/rarime/compare/2.0.2...HEAD
+[2.0.2]: https://github.com/rarimo/rarime/compare/2.0.1...2.0.2
+[2.0.1]: https://github.com/rarimo/rarime/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/rarimo/rarime/compare/1.0.2...2.0.0
 [1.0.2]: https://github.com/rarimo/rarime/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/rarimo/rarime/compare/1.0.0...1.0.1